add list of definitions for research software
This page should list definitions of the term "research software"
together with their sources and noteable uses.

At the moment, this is a proposal, and thus, only available in English.
Once accepted, a German version would need to be created as well.

This PR depends on the PR #274 for the page not to show up in the main
manu.